requirements.txt
Django==1.7.4
djangorestframework>=3.3.2,<3.4
pysolr==3.3.0
beautifulsoup4==4.3.2
marshmallow>=2.6.0,<2.7
requests==2.5.1
selenium==2.44.0
django-selenium==0.9.8
